generic json object serializer protocol to convert JSON responses to model objects
